Refactor blob creation
